+++ collect_app/src/main/java/org/odk/collect/android/widgets/SelectMultipleAutocompleteWidget.java
@@ -49,6 +49,8 @@ public class SelectMultipleAutocompleteWidget extends SelectMultiWidget implemen
 
     @Override
     protected void createLayout() {
+        readItems();
+
         if (items != null) {
             for (int i = 0; i < items.size(); i++) {
                 checkBoxes.add(createCheckBox(i));
