+++ collect_app/src/test/java/org/odk/collect/android/widgets/SelectOneWidgetTest.java
@@ -9,7 +9,7 @@ import org.robolectric.RuntimeEnvironment;
  * @author James Knight
  */
 
-public class SelectOneWidgetTest extends GeneralSelectOneWidgetTest<SelectOneWidget> {
+public class SelectOneWidgetTest extends GeneralSelectOneWidgetTest<SelectOneAbstractWidget> {
 
     @NonNull
     @Override
